Dedup code of PQFastScan and AQFastScan
