+++ faiss/CMakeLists.txt
@@ -11,7 +11,6 @@ set(FAISS_SRC
   Index.cpp
   Index2Layer.cpp
   IndexAdditiveQuantizer.cpp
-  IndexAQFastScan.cpp
   IndexBinary.cpp
   IndexBinaryFlat.cpp
   IndexBinaryFromFloat.cpp
@@ -23,9 +22,10 @@ set(FAISS_SRC
   IndexHNSW.cpp
   IndexIVF.cpp
   IndexIVFAdditiveQuantizer.cpp
-  IndexIVFAQFastScan.cpp
   IndexIVFFlat.cpp
   IndexIVFPQ.cpp
+  IndexIVFFastScan.cpp
+  IndexIVFAQFastScan.cpp
   IndexIVFPQFastScan.cpp
   IndexIVFPQR.cpp
   IndexIVFSpectralHash.cpp
@@ -34,6 +34,8 @@ set(FAISS_SRC
   IndexLattice.cpp
   IndexNSG.cpp
   IndexPQ.cpp
+  IndexFastScan.cpp
+  IndexAQFastScan.cpp
   IndexPQFastScan.cpp
   IndexPreTransform.cpp
   IndexRefine.cpp
@@ -63,8 +65,6 @@ set(FAISS_SRC
   impl/pq4_fast_scan.cpp
   impl/pq4_fast_scan_search_1.cpp
   impl/pq4_fast_scan_search_qbs.cpp
-  impl/aq4_fast_scan_search_1.cpp
-  impl/aq4_fast_scan_search_qbs.cpp
   impl/io.cpp
   impl/lattice_Zn.cpp
   impl/NNDescent.cpp
@@ -91,7 +91,6 @@ set(FAISS_HEADERS
   Index.h
   Index2Layer.h
   IndexAdditiveQuantizer.h
-  IndexAQFastScan.h
   IndexBinary.h
   IndexBinaryFlat.h
   IndexBinaryFromFloat.h
@@ -103,9 +102,10 @@ set(FAISS_HEADERS
   IndexHNSW.h
   IndexIVF.h
   IndexIVFAdditiveQuantizer.h
-  IndexIVFAQFastScan.h
   IndexIVFFlat.h
   IndexIVFPQ.h
+  IndexIVFFastScan.h
+  IndexIVFAQFastScan.h
   IndexIVFPQFastScan.h
   IndexIVFPQR.h
   IndexIVFSpectralHash.h
@@ -114,6 +114,8 @@ set(FAISS_HEADERS
   IndexNNDescent.h
   IndexNSG.h
   IndexPQ.h
+  IndexFastScan.h
+  IndexAQFastScan.h
   IndexPQFastScan.h
   IndexPreTransform.h
   IndexRefine.h
@@ -149,7 +151,6 @@ set(FAISS_HEADERS
   impl/lattice_Zn.h
   impl/platform_macros.h
   impl/pq4_fast_scan.h
-  impl/aq4_fast_scan.h
   impl/simd_result_handlers.h
   invlists/BlockInvertedLists.h
   invlists/DirectMap.h
