+++ faiss/IndexIVFAQFastScan.h
@@ -9,8 +9,8 @@
 
 #include <memory>
 
-#include <faiss/IndexIVF.h>
 #include <faiss/IndexIVFAdditiveQuantizer.h>
+#include <faiss/IndexIVFFastScan.h>
 #include <faiss/impl/AdditiveQuantizer.h>
 #include <faiss/impl/LocalSearchQuantizer.h>
 #include <faiss/utils/AlignedTable.h>
@@ -33,33 +33,10 @@ namespace faiss {
  * 13: idem, collect results in reservoir
  */
 
-struct IndexIVFAQFastScan : IndexIVF {
+struct IndexIVFAQFastScan : IndexIVFFastScan {
     using Search_type_t = AdditiveQuantizer::Search_type_t;
 
     AdditiveQuantizer* aq;
-
-    // size of the kernel
-    int bbs; // set at build time
-
-    size_t M;
-    size_t M_norm;
-    size_t nbits;
-    size_t ksub;
-
-    // M rounded up to a multiple of 2
-    size_t M2;
-
-    // search-time implementation
-    int implem = 0;
-    // skip some parts of the computation (for timing)
-    int skip = 0;
-
-    // batching factors at search time (0 = default)
-    int qbs = 0;
-    size_t qbs2 = 0;
-
-    bool rescale_norm = true;
-    int norm_scale = 1;
 
     // max number of training vectors
     size_t max_train_points;
@@ -83,12 +60,7 @@ struct IndexIVFAQFastScan : IndexIVF {
             const IndexIVFAdditiveQuantizer& orig,
             int bbs = 32);
 
-    /// orig's inverted lists (for debugging)
-    InvertedLists* orig_invlists = nullptr;
-
     void train_residual(idx_t n, const float* x) override;
-
-    void estimate_norm_scale(idx_t n, const float* x);
 
     /// same as the regular IVFAQ encoder. The codes are not reorganized by
     /// blocks a that point
@@ -99,16 +71,9 @@ struct IndexIVFAQFastScan : IndexIVF {
             uint8_t* codes,
             bool include_listno = false) const override;
 
-    void add_with_ids(idx_t n, const float* x, const idx_t* xids) override;
-
-    void search(
-            idx_t n,
-            const float* x,
-            idx_t k,
-            float* distances,
-            idx_t* labels) const override;
-
     // prepare look-up tables
+
+    bool lookup_table_is_3d() const override;
 
     void compute_LUT(
             size_t n,
@@ -118,92 +83,7 @@ struct IndexIVFAQFastScan : IndexIVF {
             AlignedTable<float>& dis_tables,
             AlignedTable<float>& biases) const;
 
-    void compute_LUT_uint8(
-            size_t n,
-            const float* x,
-            const idx_t* coarse_ids,
-            const float* coarse_dis,
-            AlignedTable<uint8_t>& dis_tables,
-            AlignedTable<uint16_t>& biases,
-            float* normalizers) const;
-
-    // internal search funcs
-
-    template <bool is_max>
-    void search_dispatch_implem(
-            idx_t n,
-            const float* x,
-            idx_t k,
-            float* distances,
-            idx_t* labels) const;
-
-    template <class C>
-    void search_implem_1(
-            idx_t n,
-            const float* x,
-            idx_t k,
-            float* distances,
-            idx_t* labels) const;
-
-    template <class C>
-    void search_implem_2(
-            idx_t n,
-            const float* x,
-            idx_t k,
-            float* distances,
-            idx_t* labels) const;
-
-    // implem 10 and 12 are not multithreaded internally, so
-    // export search stats
-    template <class C>
-    void search_implem_10(
-            idx_t n,
-            const float* x,
-            idx_t k,
-            float* distances,
-            idx_t* labels,
-            int impl,
-            size_t* ndis_out,
-            size_t* nlist_out) const;
-
-    template <class C>
-    void search_implem_12(
-            idx_t n,
-            const float* x,
-            idx_t k,
-            float* distances,
-            idx_t* labels,
-            int impl,
-            size_t* ndis_out,
-            size_t* nlist_out) const;
-
-    void reconstruct_from_offset(int64_t list_no, int64_t offset, float* recons)
-            const override;
-
-    void reconstruct_orig_invlists();
-};
-
-struct IVFAQFastScanStats {
-    uint64_t times[10];
-    uint64_t t_compute_distance_tables, t_round;
-    uint64_t t_copy_pack, t_scan, t_to_flat;
-    uint64_t reservoir_times[4];
-    double t_aq_encode;
-    double t_aq_norm_encode;
-
-    double Mcy_at(int i) {
-        return times[i] / (1000 * 1000.0);
-    }
-
-    double Mcy_reservoir_at(int i) {
-        return reservoir_times[i] / (1000 * 1000.0);
-    }
-    IVFAQFastScanStats() {
-        reset();
-    }
-    void reset() {
-        memset(this, 0, sizeof(*this));
-    }
+    void sa_decode(idx_t n, const uint8_t* bytes, float* x) const override;
 };
 
 struct IndexIVFLSQFastScan : IndexIVFAQFastScan {
@@ -241,7 +121,5 @@ struct IndexIVFRQFastScan : IndexIVFAQFastScan {
 
     ~IndexIVFRQFastScan();
 };
-
-FAISS_API extern IVFAQFastScanStats IVFAQFastScan_stats;
 
 } // namespace faiss
