+++ faiss/IndexPQFastScan.h
@@ -7,6 +7,7 @@
 
 #pragma once
 
+#include <faiss/IndexFastScan.h>
 #include <faiss/IndexPQ.h>
 #include <faiss/impl/ProductQuantizer.h>
 #include <faiss/utils/AlignedTable.h>
@@ -25,26 +26,8 @@ namespace faiss {
  * 15: no qbs with reservoir accumulator
  */
 
-struct IndexPQFastScan : Index {
+struct IndexPQFastScan : IndexFastScan {
     ProductQuantizer pq;
-
-    // implementation to select
-    int implem = 0;
-    // skip some parts of the computation (for timing)
-    int skip = 0;
-
-    // size of the kernel
-    int bbs;     // set at build time
-    int qbs = 0; // query block size 0 = use default
-
-    // packed version of the codes
-    size_t ntotal2;
-    size_t M2;
-
-    AlignedTable<uint8_t> codes;
-
-    // this is for testing purposes only (set when initialized by IndexPQ)
-    const uint8_t* orig_codes = nullptr;
 
     IndexPQFastScan(
             int d,
@@ -53,73 +36,16 @@ struct IndexPQFastScan : Index {
             MetricType metric = METRIC_L2,
             int bbs = 32);
 
-    IndexPQFastScan();
+    IndexPQFastScan() = default;
 
     /// build from an existing IndexPQ
     explicit IndexPQFastScan(const IndexPQ& orig, int bbs = 32);
 
     void train(idx_t n, const float* x) override;
-    void add(idx_t n, const float* x) override;
-    void reset() override;
-    void search(
-            idx_t n,
-            const float* x,
-            idx_t k,
-            float* distances,
-            idx_t* labels) const override;
 
-    // called by search function
-    void compute_quantized_LUT(
-            idx_t n,
-            const float* x,
-            uint8_t* lut,
-            float* normalizers) const;
+    void compute_codes(uint8_t* codes, idx_t n, const float* x) const override;
 
-    template <bool is_max>
-    void search_dispatch_implem(
-            idx_t n,
-            const float* x,
-            idx_t k,
-            float* distances,
-            idx_t* labels) const;
-
-    template <class C>
-    void search_implem_2(
-            idx_t n,
-            const float* x,
-            idx_t k,
-            float* distances,
-            idx_t* labels) const;
-
-    template <class C>
-    void search_implem_12(
-            idx_t n,
-            const float* x,
-            idx_t k,
-            float* distances,
-            idx_t* labels,
-            int impl) const;
-
-    template <class C>
-    void search_implem_14(
-            idx_t n,
-            const float* x,
-            idx_t k,
-            float* distances,
-            idx_t* labels,
-            int impl) const;
+    void compute_float_LUT(float* lut, idx_t n, const float* x) const override;
 };
-
-struct FastScanStats {
-    uint64_t t0, t1, t2, t3;
-    FastScanStats() {
-        reset();
-    }
-    void reset() {
-        memset(this, 0, sizeof(*this));
-    }
-};
-
-FAISS_API extern FastScanStats FastScan_stats;
 
 } // namespace faiss
